refactor:: 폴더 구조 변경, dto 파일명 변경
- home 폴더에 club, calendar, institution 들어가도록 변경
- getInstitutionResponseDto -> InstitutionResponseDto로 이름 변경
- createInstitutionResponseDto 파일명 오타 수정
