+++ src/app.module.ts
@@ -17,8 +17,8 @@ import { CourseReviewModule } from './course-review/course-review.module';
 import { BoardModule } from './community/board/board.module';
 import { PostModule } from './community/post/post.module';
 import { CommentModule } from './community/comment/comment.module';
-import { CalendarModule } from './calendar/calendar.module';
-import { InstitutionModule } from './institution/institution.module';
+import { CalendarModule } from './home/calendar/calendar.module';
+import { InstitutionModule } from './home/institution/institution.module';
 
 console.log(`.env.${process.env.NODE_ENV}`);
